refactor: share logging utilities
